Eliminated MSVC compilation warning.

+++ sbbs3/useredit.cpp
@@ -2,7 +2,7 @@
 
 /* Synchronet online sysop user editor */
 
-/* $Id: useredit.cpp,v 1.3 2000/10/30 12:22:03 rswindell Exp $ */
+/* $Id: useredit.cpp,v 1.4 2000/10/31 04:13:44 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -646,7 +646,7 @@ int sbbs_t::searchup(char *search,int usernum)
 	flen=filelength(file);
 	lseek(file,(long)((long)usernum*U_LEN),0);
 
-	while(i*U_LEN<=flen) {
+	while((i*U_LEN)<=(ulong)flen) {
 		count=0;
 		while(count<LOOP_NODEDAB
 			&& lock(file,(long)((long)(i-1)*U_LEN),U_LEN)==-1) {
